fix: rename shared/tests to shared/test-utils
