+++ www/app/App.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Router from 'next/router';
 import { render } from '@testing-library/react';
 import { App } from './App';
-import { AppTreeWrapper } from '../shared/tests';
+import { AppTreeWrapper } from '../shared/test-utils';
 import { trackPageViews } from '../shared/utils/google-analytics';
 
 jest.mock('../shared/utils/google-analytics', () => ({ trackPageViews: jest.fn(() => jest.fn()) }));
